move tprint and is_sequence to better locations
