+++ qcodes/tests/helpers/test_issequenceof.py
@@ -2,7 +2,7 @@ from typing import Any, List, Tuple
 
 import pytest
 
-from qcodes.utils.helpers import is_sequence_of
+from qcodes.parameters.sequence_helpers import is_sequence_of
 
 simple_good: List[Tuple[Any, ...]] = [
     # empty lists pass without even checking that we provided a
